+++ lib/src/editor/raw_editor/config/events/standard_char_shortcuts/standard_shortcut_events.dart
@@ -1,0 +1,23 @@
+import '../events.dart';
+
+/// These all the common CharacterShortcutEvents that are implemented
+/// by the package and correspond with markdown syntax
+final standardCharactersShortcutEvents =
+    List<CharacterShortcutEvent>.unmodifiable(<CharacterShortcutEvent>[
+  formatAsterisksToItalic,
+  formatStrikeToStrikethrough,
+  formatCodeCharToInlineCode,
+  formatDoubleAsterisksToBold,
+  formatDoubleUnderscoresToBold,
+]);
+
+/// These all the common SpaceShortcutEvent that are implemented
+/// by the package and correspond with markdown syntax
+final standardSpaceShorcutEvents =
+    List<SpaceShortcutEvent>.unmodifiable(<SpaceShortcutEvent>[
+  formatOrderedNumberToList,
+  formatHyphenToBulletList,
+  formatHeaderToHeaderStyle,
+  formatHeader2ToHeaderStyle,
+  formatHeader3ToHeaderStyle,
+]);