+++ lib/src/editor/raw_editor/raw_editor_state.dart
@@ -532,14 +532,14 @@ class QuillRawEditorState extends EditorState
         data: _styles!,
         child: QuillKeyboardServiceWidget(
           actions: _actions,
+          characterEvents: widget.configurations.characterShortcutEvents,
+          spaceEvents: widget.configurations.spaceShortcutEvents,
           constraints: constraints,
           focusNode: widget.configurations.focusNode,
           controller: controller,
           readOnly: widget.configurations.readOnly,
           enableAlwaysIndentOnTab:
               widget.configurations.enableAlwaysIndentOnTab,
-          enableMdConversion:
-              widget.configurations.enableMarkdownStyleConversion,
           customShortcuts: widget.configurations.customShortcuts,
           customActions: widget.configurations.customActions,
           child: child,
@@ -719,7 +719,8 @@ class QuillRawEditorState extends EditorState
         widget.configurations.enableInteractiveSelection,
         _hasFocus,
         MediaQuery.devicePixelRatioOf(context),
-        _cursorCont);
+        _cursorCont,
+        _styles!.inlineCode!);
     return editableTextLine;
   }